refactor(web): isolate focused moment scrolling
Move focused-moment navigation into a dedicated hook and add unit coverage for label precedence, semantic-moment fallback, and chunk-loading behavior.
